Centralize team label normalization
